+++ examples/permissioned-token-freezing/test/asset-txfer-test.js
@@ -115,14 +115,7 @@ describe('Test for transferring asset using custom logic', function () {
     console.log('Escrow Address: ', escrowAddress);
 
     // fund escrow with some minimum balance first
-    runtime.executeTx({
-      type: types.TransactionType.TransferAlgo,
-      sign: types.SignType.SecretKey,
-      fromAccount: master.account,
-      toAccountAddr: escrowAddress,
-      amountMicroAlgos: minBalance,
-      payFlags: {}
-    });
+    runtime.fundLsig(master.account, escrowAddress, minBalance);
 
     /** Update clawback address to escrow + Locking the manager and freeze address **/
     const assetModFields = {