+++ packages/runtime/src/parser/parser.ts
@@ -6,16 +6,17 @@ import {
   AppGlobalPut, AppLocalDel, AppLocalGet, AppLocalGetEx, AppLocalPut,
   AppOptedIn, Arg, Assert, Balance, BitwiseAnd, BitwiseNot, BitwiseOr,
   BitwiseXor, Branch, BranchIfNotZero, BranchIfNotZerov4, BranchIfZero, BranchIfZerov4, Branchv4,
-  Btoi, Byte, ByteAdd, Bytec, Bytecblock, Callsub,
+  Btoi, Byte, ByteAdd, ByteBitwiseAnd, ByteBitwiseInvert, ByteBitwiseOr,
+  ByteBitwiseXor, Bytec, Bytecblock, ByteDiv, ByteEqualTo, ByteGreaterThanEqualTo, ByteGreatorThan,
+  ByteLessThan, ByteLessThanEqualTo, ByteMod, ByteMul, ByteNotEqualTo, ByteSub,
+  ByteZero, Callsub,
   Concat, Dig, Div, DivModw, Dup, Dup2, Ed25519verify,
-  EqualTo, Err, Exp, Expw, GetAssetDef, GetAssetHolding, GetBit, GetByte, Gload,
-  Gloads, Global, GreaterThan, GreaterThanEqualTo, Gtxn,
-  Gtxna, Gtxns, Gtxnsa, Int, Intc, Intcblock, Itob,
-  Keccak256, Label, Len, LessThan, LessThanEqualTo,
-  Load, MinBalance, Mod, Mul, Mulw, Not, NotEqualTo,
-  Or, Pop, Pragma, PushBytes, PushInt, Retsub,
-  Return, Select, SetBit, SetByte, Sha256, Sha512_256,
-  Shl, Shr, Sqrt, Store, Sub, Substring, Substring3, Swap, Txn, Txna
+  EqualTo, Err, Exp, Expw, GetAssetDef, GetAssetHolding, GetBit, GetByte, Gload, Gloads, Global, GreaterThan,
+  GreaterThanEqualTo, Gtxn, Gtxna, Gtxns, Gtxnsa, Int, Intc, Intcblock, Itob,
+  Keccak256, Label, Len, LessThan, LessThanEqualTo, Load, MinBalance, Mod,
+  Mul, Mulw, Not, NotEqualTo, Or, Pop, Pragma, PushBytes, PushInt, Retsub,
+  Return, Select, SetBit, SetByte, Sha256,
+  Sha512_256, Shl, Shr, Sqrt, Store, Sub, Substring, Substring3, Swap, Txn, Txna
 } from "../interpreter/opcode-list";
 import { LogicSigMaxCost, LogicSigMaxSize, MaxAppProgramCost, MaxAppProgramLen, OpGasCost } from "../lib/constants";
 import { assertLen } from "../lib/parsing";
@@ -171,6 +172,21 @@ opCodeMap[4] = {
   bz: BranchIfZerov4,
   // byteslice arithmetic ops
   'b+': ByteAdd,
+  'b-': ByteSub,
+  'b*': ByteMul,
+  'b/': ByteDiv,
+  'b%': ByteMod,
+  'b<': ByteLessThan,
+  'b>': ByteGreatorThan,
+  'b<=': ByteLessThanEqualTo,
+  'b>=': ByteGreaterThanEqualTo,
+  'b==': ByteEqualTo,
+  'b!=': ByteNotEqualTo,
+  'b|': ByteBitwiseOr,
+  'b&': ByteBitwiseAnd,
+  'b^': ByteBitwiseXor,
+  'b~': ByteBitwiseInvert,
+  bzero: ByteZero,
 
   divmodw: DivModw,
   exp: Exp,